pyproject.toml
[project]
name = "ai-engineering-hampus-donnersten-de24"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.122.0",
    "google-genai>=1.51.0",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pydantic-ai>=1.24.0",
    "pydantic>=2.12.4",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
    "python-dotenv>=1.2.1",
    "duckdb>=1.4.2",
    "streamlit>=1.51.0",
    "lancedb>=0.25.3",
    "google-generativeai>=0.8.5",
]
